+++ pkg/artifactory/resource/repository/remote/resource_artifactory_remote_bower_repository.go
@@ -9,7 +9,7 @@ import (
 )
 
 type BowerRemoteRepo struct {
-	RepositoryBaseParams
+	RepositoryRemoteBaseParams
 	RepositoryVcsParams
 	BowerRegistryUrl string `json:"bowerRegistryUrl"`
 }
@@ -30,9 +30,9 @@ func ResourceArtifactoryRemoteBowerRepository() *schema.Resource {
 	var unpackBowerRemoteRepo = func(s *schema.ResourceData) (interface{}, string, error) {
 		d := &util.ResourceData{ResourceData: s}
 		repo := BowerRemoteRepo{
-			RepositoryBaseParams: UnpackBaseRemoteRepo(s, packageType),
-			RepositoryVcsParams:  UnpackVcsRemoteRepo(s),
-			BowerRegistryUrl:     d.GetString("bower_registry_url", false),
+			RepositoryRemoteBaseParams: UnpackBaseRemoteRepo(s, packageType),
+			RepositoryVcsParams:        UnpackVcsRemoteRepo(s),
+			BowerRegistryUrl:           d.GetString("bower_registry_url", false),
 		}
 		return repo, repo.Id(), nil
 	}
@@ -44,7 +44,7 @@ func ResourceArtifactoryRemoteBowerRepository() *schema.Resource {
 		}
 
 		return &BowerRemoteRepo{
-			RepositoryBaseParams: RepositoryBaseParams{
+			RepositoryRemoteBaseParams: RepositoryRemoteBaseParams{
 				Rclass:              "remote",
 				PackageType:         packageType,
 				RemoteRepoLayoutRef: repoLayout.(string),