+++ pkg/artifactory/resource/repository/remote/resource_artifactory_remote_cocoapods_repository.go
@@ -9,7 +9,7 @@ import (
 )
 
 type CocoapodsRemoteRepo struct {
-	RepositoryBaseParams
+	RepositoryRemoteBaseParams
 	RepositoryVcsParams
 	PodsSpecsRepoUrl string `json:"podsSpecsRepoUrl"`
 }
@@ -30,9 +30,9 @@ func ResourceArtifactoryRemoteCocoapodsRepository() *schema.Resource {
 	var unpackCocoapodsRemoteRepo = func(s *schema.ResourceData) (interface{}, string, error) {
 		d := &util.ResourceData{ResourceData: s}
 		repo := CocoapodsRemoteRepo{
-			RepositoryBaseParams: UnpackBaseRemoteRepo(s, packageType),
-			RepositoryVcsParams:  UnpackVcsRemoteRepo(s),
-			PodsSpecsRepoUrl:     d.GetString("pods_specs_repo_url", false),
+			RepositoryRemoteBaseParams: UnpackBaseRemoteRepo(s, packageType),
+			RepositoryVcsParams:        UnpackVcsRemoteRepo(s),
+			PodsSpecsRepoUrl:           d.GetString("pods_specs_repo_url", false),
 		}
 		return repo, repo.Id(), nil
 	}
@@ -44,7 +44,7 @@ func ResourceArtifactoryRemoteCocoapodsRepository() *schema.Resource {
 		}
 
 		return &CocoapodsRemoteRepo{
-			RepositoryBaseParams: RepositoryBaseParams{
+			RepositoryRemoteBaseParams: RepositoryRemoteBaseParams{
 				Rclass:              "remote",
 				PackageType:         packageType,
 				RemoteRepoLayoutRef: repoLayout.(string),